5-1 homework fix1
