+++ src/presenter/trip-presenter.js
@@ -5,7 +5,7 @@ import PointPresenter from './point-presenter.js';
 import { updateItem } from '../utils/common.js';
 import ListSortView from '../view/list-sort-view.js';
 import { SortType } from '../const.js';
-import { sortedPoints } from '../utils/filter.js';
+import { sortPoints } from '../utils/task';
 
 
 const mainEventsElement = document.querySelector('.trip-events');
@@ -55,20 +55,18 @@ export default class TripPresenter {
   #sortPoints(sortType) {
     switch (sortType) {
       case SortType.DAY:
-        this.#listPoints = sortedPoints(this.#listPoints, SortType.DAY);
+        this.#listPoints = sortPoints(this.#listPoints, SortType.DAY);
         break;
       case SortType.TIME:
-        this.#listPoints = sortedPoints(this.#listPoints, SortType.TIME);
+        this.#listPoints = sortPoints(this.#listPoints, SortType.TIME);
         break;
       case SortType.PRICE:
-        this.#listPoints = sortedPoints(this.#listPoints, SortType.PRICE);
+        this.#listPoints = sortPoints(this.#listPoints, SortType.PRICE);
         break;
       case SortType.OFFERS:
-        this.#listPoints = sortedPoints(this.#listPoints, SortType.OFFERS);
+        this.#listPoints = sortPoints(this.#listPoints, SortType.OFFERS);
         break;
       default:
-        // 3. А когда пользователь захочет "вернуть всё, как было",
-        // мы просто запишем в _boardTasks исходный массив
         this.#listPoints = [...this.#sourcedListPoints];
     }
