+++ src/utils/task.js
@@ -1,4 +1,5 @@
 import dayjs from 'dayjs';
+import { SortType } from '../const.js';
 
 const DATE_FORMAT = 'MMM DD';
 
@@ -44,11 +45,37 @@ function sortPointDown(pointA, pointB) {
   return weight ?? dayjs(pointB.dueDate).diff(dayjs(pointA.dueDate));
 }
 
+const sortPoints = (points, sortType) => {
+  switch (sortType) {
+    case SortType.DAY:
+      return points.sort((pointA, pointB) => dayjs(pointA.dateFrom).diff(dayjs(pointB.dateFrom)));
+    case SortType.TIME:
+      return points.sort((pointA, pointB) =>
+        dayjs(pointA.dateTo).diff(pointA.dateFrom) - dayjs(pointB.dateTo).diff(pointB.dateFrom));
+    case SortType.OFFERS:
+      return points.sort((pointA, pointB) => pointA.offers.length - pointB.offers.length);
+    case SortType.PRICE:
+      return points.sort((pointA, pointB) => pointA.basePrice - pointB.basePrice);
+    default:
+      return points;
+  }
+};
+
+const sort = {
+  [SortType.DAY]: (points) => points.sort((from, to) => dayjs(from.dateFrom).diff(dayjs(to.dateFrom))),
+  [SortType.EVENT]: (points) => points,
+  [SortType.TIME]: (points) => points,
+  [SortType.PRICE]: (points) => points.sort((from, to) => to.basePrice - from.basePrice),
+  [SortType.OFFERS]: (points) => points,
+};
+
 export {
   humanizeTaskDueDate,
   isTaskExpired,
   isTaskExpiringToday,
   sortPointDown,
   sortPointUp,
-  getWeightForNullDate
+  getWeightForNullDate,
+  sortPoints,
+  sort
 };
